Ability to switch between multiple queues
With the existing implementation, the machinery worker only listens to
a single queue and fetch messages from it. This PR is to make it
dynamic and add ability to switch between multiple queues, if needed.

One use case being when we've multiple different queues and we want
to fetch messages from each one of them in round-robin. The changes
however are done in such a way that it makes the queue dyamic and
devs could plugin their own custom logic, if needed.